the rays are casted away but for the north

+++ headers/struct.h
@@ -6,7 +6,7 @@
 /*   By: Ma3ert <yait-iaz@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/12 12:16:30 by Ma3ert            #+#    #+#             */
-/*   Updated: 2022/08/23 16:03:36 by Ma3ert           ###   ########.fr       */
+/*   Updated: 2022/08/24 15:02:31 by Ma3ert           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,8 +75,8 @@ typedef struct s_ray
 	double		ybound;		// the y cord of the horizontal intersection
 	int			v_hit;		// the type of the cell for the vertical intersect
 	int			h_hit;		// the type of the cell for the horizontal intersect
-	int			v_distance;	// the distance between the position of the player
-	int			h_distance;	// and the intesection
+	double		v_distance;	// the distance between the position of the player
+	double		h_distance;	// and the intesection
 	double		x_step;		// step to get to the next x intersection 
 	double		y_step;		// step to get to the next y intersection
 	int			quadrant;	// in which quadrant the ray is casted
@@ -86,7 +86,6 @@ typedef struct s_ray
 	int			ycell_h;	// the y cord of cell for the horizontal interesection
 	t_position	*player;	// info about the position of the player
 }				t_ray;
-
 
 typedef struct s_table
 {
